pyproject.toml
[project]
name = "social-media-posts-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==24.1.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "asttokens==3.0.0",
    "audioop-lts==0.2.2",
    "brotli==1.1.0",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "decorator==5.2.1",
    "executing==2.2.1",
    "fastapi==0.119.1",
    "ffmpy==0.6.3",
    "filelock==3.20.0",
    "filetype==1.2.0",
    "fsspec==2025.9.0",
    "google-ai-generativelanguage==0.9.0",
    "google-api-core==2.26.0",
    "google-api-python-client==2.185.0",
    "google-auth==2.41.1",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.71.0",
    "gradio==5.49.1",
    "gradio-client==1.13.3",
    "groovy==0.1.2",
    "grpcio==1.75.1",
    "grpcio-status==1.75.1",
    "h11==0.16.0",
    "hf-xet==1.1.10",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httpx==0.28.1",
    "huggingface-hub==0.35.3",
    "idna==3.11",
    "ipython==9.6.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-core==1.0.0",
    "langchain-google-genai==3.0.0",
    "langgraph==1.0.1",
    "langgraph-checkpoint==3.0.0",
    "langgraph-prebuilt==1.0.1",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.37",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "numpy==2.3.4",
    "oauthlib==3.3.1",
    "orjson==3.11.3",
    "ormsgpack==1.11.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "prompt-toolkit==3.0.52",
    "proto-plus==1.26.1",
    "protobuf==6.33.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.10",
    "pydantic-core==2.33.2",
    "pydub==0.25.1",
    "pygments==2.19.2",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rsa==4.9.1",
    "ruff==0.14.1",
    "safehttpx==0.1.6",
    "semantic-version==2.10.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "stack-data==0.6.3",
    "starlette==0.48.0",
    "tenacity==9.1.2",
    "tomlkit==0.13.3",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "wcwidth==0.2.14",
    "websockets==15.0.1",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
]
